requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
